refactor: move dispatch source generation to util
