+++ src/types/ActionTypes.ts
@@ -106,6 +106,9 @@ export enum AT {
     FETCH_TRANSCRIPT_VALIDATION_ASYNC = 'FETCH_TRANSCRIPT_VALIDATION_ASYNC',
     FETCH_TRANSCRIPT_VALIDATION_FULFILLED = 'FETCH_TRANSCRIPT_VALIDATION_FULFILLED',
 
+    REGENERATE_DISPATCH_DIALOGS_ASYNC = 'REGENERATE_DISPATCH_DIALOGS_ASYNC',
+    REGENERATE_DISPATCH_DIALOGS_FULFILLED = 'REGENERATE_DISPATCH_DIALOGS_FULFILLED',
+
     // CreateActions
     COPY_APPLICATION_ASYNC = 'COPY_APPLICATIONS_ASYNC',
     COPY_APPLICATION_FULFILLED = 'COPY_APPLICATIONS_FULFILLED',
